pyproject.toml
[project]
name = "chip8"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "glfw>=2.9.0",
    "numpy>=2.2.6",
    "pyopengl>=3.1.9",
]

[dependency-groups]
dev = [
    "mypy>=1.16.0",
    "ruff>=0.11.12",
]
